Finished code cleanup, Fixed broken unit tests
