+++ src/test/java/com/monumental/services/integrationtest/MonumentServiceIntegrationTests.java
@@ -2122,6 +2122,370 @@ public class MonumentServiceIntegrationTests {
         assertEquals(12, this.tagRepository.getAllByMonumentId(result.getId()).size());
     }
 
+    /* createMonumentReferences Tests */
+
+    @Test
+    public void testMonumentService_createMonumentReferences_NullReferenceUrls_NonNullMonument() {
+        Monument monument = new Monument();
+        assertNull(this.monumentService.createMonumentReferences(null, monument));
+    }
+
+    @Test
+    public void testMonumentService_createMonumentReferences_NonNullReferenceUrls_NullMonument() {
+        List<String> referenceUrls = new ArrayList<>();
+        assertNull(this.monumentService.createMonumentReferences(referenceUrls, null));
+    }
+
+    @Test
+    public void testMonumentService_createMonumentReferences_NullReferenceUrls_NullMonument() {
+        assertNull(this.monumentService.createMonumentReferences(null, null));
+    }
+
+    @Test
+    public void testMonumentService_createMonumentReferences_EmptyReferenceUrls() {
+        List<String> referenceUrls = new ArrayList<>();
+        Monument monument = new Monument();
+        monument = this.monumentRepository.save(monument);
+
+        List<Reference> result = this.monumentService.createMonumentReferences(referenceUrls, monument);
+
+        assertEquals(0, result.size());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentReferences_OneNullReferenceUrl() {
+        List<String> referenceUrls = new ArrayList<>();
+        referenceUrls.add(null);
+
+        Monument monument = new Monument();
+        monument = this.monumentRepository.save(monument);
+
+        List<Reference> result = this.monumentService.createMonumentReferences(referenceUrls, monument);
+
+        assertEquals(0, result.size());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentReferences_OneEmptyReferenceUrl() {
+        List<String> referenceUrls = new ArrayList<>();
+        referenceUrls.add("");
+
+        Monument monument = new Monument();
+        monument = this.monumentRepository.save(monument);
+
+        List<Reference> result = this.monumentService.createMonumentReferences(referenceUrls, monument);
+
+        assertEquals(0, result.size());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentReferences_OneNullAndOneEmptyReferenceUrl() {
+        List<String> referenceUrls = new ArrayList<>();
+        referenceUrls.add(null);
+        referenceUrls.add("");
+
+        Monument monument = new Monument();
+        monument = this.monumentRepository.save(monument);
+
+        List<Reference> result = this.monumentService.createMonumentReferences(referenceUrls, monument);
+
+        assertEquals(0, result.size());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentReferences_OneReferenceUrl() {
+        List<String> referenceUrls = new ArrayList<>();
+        referenceUrls.add("test");
+
+        Monument monument = new Monument();
+        monument.setTitle("Monument");
+        monument = this.monumentRepository.save(monument);
+
+        List<Reference> result = this.monumentService.createMonumentReferences(referenceUrls, monument);
+
+        assertEquals(1, result.size());
+
+        Reference reference = result.get(0);
+        assertEquals("test", reference.getUrl());
+        assertEquals("Monument", reference.getMonument().getTitle());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentReferences_ThreeReferenceUrlsWithOneNullAndOneEmpty() {
+        List<String> referenceUrls = new ArrayList<>();
+        referenceUrls.add("test1");
+        referenceUrls.add("test2");
+        referenceUrls.add("test3");
+        referenceUrls.add(null);
+        referenceUrls.add("");
+
+        Monument monument = new Monument();
+        monument.setTitle("Monument");
+        monument = this.monumentRepository.save(monument);
+
+        List<Reference> result = this.monumentService.createMonumentReferences(referenceUrls, monument);
+
+        assertEquals(3, result.size());
+
+        Reference reference1 = result.get(0);
+        assertEquals("test1", reference1.getUrl());
+        assertEquals("Monument", reference1.getMonument().getTitle());
+
+        Reference reference2 = result.get(1);
+        assertEquals("test2", reference2.getUrl());
+        assertEquals("Monument", reference2.getMonument().getTitle());
+
+        Reference reference3 = result.get(2);
+        assertEquals("test3", reference3.getUrl());
+        assertEquals("Monument", reference3.getMonument().getTitle());
+    }
+
+    /* createMonumentImages Tests */
+
+    @Test
+    public void testMonumentService_createMonumentImages_NullImagesUrls_NonNullMonument() {
+        Monument monument = new Monument();
+        assertNull(this.monumentService.createMonumentImages(null, monument));
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_NonNullImagesUrls_NullMonument() {
+        List<String> imageUrls = new ArrayList<>();
+        assertNull(this.monumentService.createMonumentImages(imageUrls, null));
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_NullImageUrls_NullMonument() {
+        assertNull(this.monumentService.createMonumentImages(null, null));
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_EmptyImageUrls() {
+        List<String> imageUrls = new ArrayList<>();
+        Monument monument = new Monument();
+        monument = this.monumentRepository.save(monument);
+
+        List<Image> result = this.monumentService.createMonumentImages(imageUrls, monument);
+
+        assertEquals(0, result.size());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_OneNullImageUrl() {
+        List<String> imageUrls = new ArrayList<>();
+        imageUrls.add(null);
+
+        Monument monument = new Monument();
+        monument = this.monumentRepository.save(monument);
+
+        List<Image> result = this.monumentService.createMonumentImages(imageUrls, monument);
+
+        assertEquals(0, result.size());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_OneEmptyImageUrl() {
+        List<String> imageUrls = new ArrayList<>();
+        imageUrls.add("");
+
+        Monument monument = new Monument();
+        monument = this.monumentRepository.save(monument);
+
+        List<Image> result = this.monumentService.createMonumentImages(imageUrls, monument);
+
+        assertEquals(0, result.size());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_OneNullAndOneEmptyImageUrl() {
+        List<String> imageUrls = new ArrayList<>();
+        imageUrls.add(null);
+        imageUrls.add("");
+
+        Monument monument = new Monument();
+        monument = this.monumentRepository.save(monument);
+
+        List<Image> result = this.monumentService.createMonumentImages(imageUrls, monument);
+
+        assertEquals(0, result.size());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_OneImageUrl() {
+        List<String> imageUrls = new ArrayList<>();
+        imageUrls.add("test");
+
+        Monument monument = new Monument();
+        monument.setTitle("Monument");
+        monument = this.monumentRepository.save(monument);
+
+        List<Image> result = this.monumentService.createMonumentImages(imageUrls, monument);
+
+        assertEquals(1, result.size());
+
+        Image image = result.get(0);
+        assertEquals("test", image.getUrl());
+        assertTrue(image.getIsPrimary());
+        assertEquals("Monument", image.getMonument().getTitle());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_OneImageUrl_MonumentAlreadyHasImages_NoPrimary() {
+        List<String> imageUrls = new ArrayList<>();
+        imageUrls.add("test1");
+
+        Monument monument = new Monument();
+        monument.setTitle("Monument");
+        monument = this.monumentRepository.save(monument);
+
+        List<Image> monumentImages = new ArrayList<>();
+        monumentImages.add(new Image("test2", false));
+
+        monument.setImages(monumentImages);
+
+        List<Image> result = this.monumentService.createMonumentImages(imageUrls, monument);
+
+        assertEquals(1, result.size());
+
+        Image image = result.get(0);
+        assertEquals("test1", image.getUrl());
+        assertTrue(image.getIsPrimary());
+        assertEquals("Monument", image.getMonument().getTitle());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_OneImageUrl_MonumentAlreadyHasPrimaryImage() {
+        List<String> imageUrls = new ArrayList<>();
+        imageUrls.add("test1");
+
+        Monument monument = new Monument();
+        monument.setTitle("Monument");
+        monument = this.monumentRepository.save(monument);
+
+        List<Image> monumentImages = new ArrayList<>();
+        monumentImages.add(new Image("test2", true));
+
+        monument.setImages(monumentImages);
+
+        List<Image> result = this.monumentService.createMonumentImages(imageUrls, monument);
+
+        assertEquals(1, result.size());
+
+        Image image = result.get(0);
+        assertEquals("test1", image.getUrl());
+        assertFalse(image.getIsPrimary());
+        assertEquals("Monument", image.getMonument().getTitle());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_ThreeImageUrlsWithOneNullAndOneEmpty() {
+        List<String> imageUrls = new ArrayList<>();
+        imageUrls.add("test1");
+        imageUrls.add("test2");
+        imageUrls.add("test3");
+        imageUrls.add(null);
+        imageUrls.add("");
+
+        Monument monument = new Monument();
+        monument.setTitle("Monument");
+        monument = this.monumentRepository.save(monument);
+
+        List<Image> result = this.monumentService.createMonumentImages(imageUrls, monument);
+
+        assertEquals(3, result.size());
+
+        Image image1 = result.get(0);
+        assertEquals("test1", image1.getUrl());
+        assertTrue(image1.getIsPrimary());
+        assertEquals("Monument", image1.getMonument().getTitle());
+
+        Image image2 = result.get(1);
+        assertEquals("test2", image2.getUrl());
+        assertFalse(image2.getIsPrimary());
+        assertEquals("Monument", image2.getMonument().getTitle());
+
+        Image image3 = result.get(2);
+        assertEquals("test3", image3.getUrl());
+        assertFalse(image3.getIsPrimary());
+        assertEquals("Monument", image3.getMonument().getTitle());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_ThreeImageUrlsWithOneNullAndOneEmpty_MonumentAlreadyHasImages_NoPrimary() {
+        List<String> imageUrls = new ArrayList<>();
+        imageUrls.add("test1");
+        imageUrls.add("test2");
+        imageUrls.add("test3");
+        imageUrls.add(null);
+        imageUrls.add("");
+
+        Monument monument = new Monument();
+        monument.setTitle("Monument");
+        monument = this.monumentRepository.save(monument);
+
+        List<Image> monumentImages = new ArrayList<>();
+        monumentImages.add(new Image("test4", false));
+
+        monument.setImages(monumentImages);
+
+        List<Image> result = this.monumentService.createMonumentImages(imageUrls, monument);
+
+        assertEquals(3, result.size());
+
+        Image image1 = result.get(0);
+        assertEquals("test1", image1.getUrl());
+        assertTrue(image1.getIsPrimary());
+        assertEquals("Monument", image1.getMonument().getTitle());
+
+        Image image2 = result.get(1);
+        assertEquals("test2", image2.getUrl());
+        assertFalse(image2.getIsPrimary());
+        assertEquals("Monument", image2.getMonument().getTitle());
+
+        Image image3 = result.get(2);
+        assertEquals("test3", image3.getUrl());
+        assertFalse(image3.getIsPrimary());
+        assertEquals("Monument", image3.getMonument().getTitle());
+    }
+
+    @Test
+    public void testMonumentService_createMonumentImages_ThreeImageUrlsWithOneNullAndOneEmpty_MonumentAlreadyHasPrimaryImage() {
+        List<String> imageUrls = new ArrayList<>();
+        imageUrls.add("test1");
+        imageUrls.add("test2");
+        imageUrls.add("test3");
+        imageUrls.add(null);
+        imageUrls.add("");
+
+        Monument monument = new Monument();
+        monument.setTitle("Monument");
+        monument = this.monumentRepository.save(monument);
+
+        List<Image> monumentImages = new ArrayList<>();
+        monumentImages.add(new Image("test4", true));
+
+        monument.setImages(monumentImages);
+
+        List<Image> result = this.monumentService.createMonumentImages(imageUrls, monument);
+
+        assertEquals(3, result.size());
+
+        Image image1 = result.get(0);
+        assertEquals("test1", image1.getUrl());
+        assertFalse(image1.getIsPrimary());
+        assertEquals("Monument", image1.getMonument().getTitle());
+
+        Image image2 = result.get(1);
+        assertEquals("test2", image2.getUrl());
+        assertFalse(image2.getIsPrimary());
+        assertEquals("Monument", image2.getMonument().getTitle());
+
+        Image image3 = result.get(2);
+        assertEquals("test3", image3.getUrl());
+        assertFalse(image3.getIsPrimary());
+        assertEquals("Monument", image3.getMonument().getTitle());
+    }
+
     /* updateMonumentTags Tests */
 
     @Test
@@ -2981,17 +3345,18 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(3, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 1) {
-                assertEquals("New Reference URL 1", reference.getUrl());
-            }
-            else if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 1"));
+        assertFalse(referenceUrls.contains("Reference URL 1"));
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
     }
 
     @Test
@@ -3071,14 +3436,15 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
     }
 
     @Test
@@ -3165,14 +3531,15 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(3, monument.getImages().size());
         assertTrue(monument.getImages().get(0).getIsPrimary());
@@ -3280,19 +3647,20 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(5, monument.getImages().size());
 
         for (Image image : monument.getImages()) {
-            if (image.getId() == 1) {
+            if (image.getUrl().equals("Image URL 1")) {
                 assertTrue(image.getIsPrimary());
             }
             else {
@@ -3404,19 +3772,20 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(5, monument.getImages().size());
 
         for (Image image : monument.getImages()) {
-            if (image.getId() == 2) {
+            if (image.getUrl().equals("Image URL 2")) {
                 assertTrue(image.getIsPrimary());
             }
             else {
@@ -3532,19 +3901,20 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(4, monument.getImages().size());
 
         for (Image image : monument.getImages()) {
-            if (image.getId() == 2) {
+            if (image.getUrl().equals("Image URL 2")) {
                 assertTrue(image.getIsPrimary());
             }
             else {
@@ -3660,14 +4030,15 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(4, monument.getImages().size());
 
@@ -3796,14 +4167,15 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(4, monument.getImages().size());
 
@@ -3944,14 +4316,15 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(4, monument.getImages().size());
 
@@ -4091,14 +4464,15 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(4, monument.getImages().size());
 
@@ -4245,14 +4619,15 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(4, monument.getImages().size());
 
@@ -4404,14 +4779,15 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(4, monument.getImages().size());
 
@@ -4562,14 +4938,15 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, monument.getReferences().size());
 
+        List<String> referenceUrls = new ArrayList<>();
         for (Reference reference : monument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(4, monument.getImages().size());
 
@@ -4722,14 +5099,15 @@ public class MonumentServiceIntegrationTests {
 
         assertEquals(2, updatedMonument.getReferences().size());
 
-        for (Reference reference : updatedMonument.getReferences()) {
-            if (reference.getId() == 2) {
-                assertEquals("New Reference URL 2", reference.getUrl());
-            }
-            else {
-                assertEquals("Reference URL 3", reference.getUrl());
-            }
+        List<String> referenceUrls = new ArrayList<>();
+        for (Reference reference : monument.getReferences()) {
+            referenceUrls.add(reference.getUrl());
         }
+
+        assertTrue(referenceUrls.contains("New Reference URL 2"));
+        assertFalse(referenceUrls.contains("Reference URL 2"));
+
+        assertTrue(referenceUrls.contains("Reference URL 3"));
 
         assertEquals(2, this.referenceRepository.getAllByMonumentId(updatedMonument.getId()).size());
